mkcurl.h: fix build on Windows x86
Sizeof(size_t) < sizeof(uint64_t) on Windows x86. Fix by requiring
the size to be simply a size_t. While there, fix all the other cases
where we return a size_t type into a uint64_t. Not doing that is
a thing that will causes us headaches in the future.

Change the function names so that it will not compile, even with
lower levels of warnings, and will surely notice.