mkcurl.h: only return to caller the certificates
I have noticed that fields other from the certificates MAY contain
binary data, and I'd like the certificates to be very easy to handle
without the complication of accessing binary data.

So, remove that extra data and only provide certificates.

This breaks the contract with users, but we've already said we need
to bump to v0.7.0, hence also this breaking change is covered.