cmake: on Windows compile unit-tests.cpp with /EHs
We use synchronous exceptions in unit-tests.cpp to make sure
that we abort when required without actually aborting.

To this end, we need to compile unit-tests.cpp with /EHs that
assumes ordinary exeptions for all functions.

The default /EHsc assumes that `extern C` functions don't throw.

This is fine in the common case, but not when testing.