Rewrite to have pure C++ API
The general idea of having a simple C API that is very easily wrapped
from JNI/ObjectiveC/Go is a very good idea and I'll continue using
this pattern. However, there are two additional considerations:

1. this is a very low level library that it's unlikely to be exposed
to the MK user

2. even if we want to do that, the C wrapper to the C++ API should
probably go directly in MK and heavily use macros to more easily
automate the task of wrapping

By having only the C++ API here, we enable dependent pieces of code
to more easily interact with this code, and we make this header-only
library slimmer. This increases the code velocity (i.e. our ability
of quickly changing the code and quickly understanding it in the
future).