mkcurl: allow reusing easy handle
Closes #34

This should allow changes in mkcollector such that we can reuse an
existing cURL handle when submitting a batch of reports.

In turn, this means we can cut the time require for uploading reports
because wen can reuse the existing connection.

I'd like this to be included into the next OONI mobile release and
to do this there is a number of steps of which this is the first.

Related to https://github.com/ooni/probe/issues/861