mkcurl.hpp: inline namespace for public symbols
This allows us to have more than one version of mkcurl inside of
android-libs and/or mkall-ios for testing pursposes. Currently mkcurl
is part of MK, so testing is much slower then that as we need to
vendor mkcurl into MK, do a full build of MK, and the use that with
the apps. With this change, we can be quicker because we can vendor
the new version of mkcurl and test it out directly. If that's OK
then we can go for vendoring into MK and doing a full build.

Xref: https://github.com/measurement-kit/measurement-kit/issues/1867

See also: https://github.com/measurement-kit/mkcollector/pull/25#discussion_r305323787